Cli Migration V3
refactor of cli migration to be faster and more accurate. Previously, the tool relied on two png images (old_tiles.png vs new_tiles.png) and using image comparison, figured out what the new bindings were.

In the refactor, it uses a new_tiles.xml to get the binding of old to new more directly. Image software is still used to display forgotten tiles to the user. This method is more elegant, accurate, and fast!