Scroll Adder 1.2
- Fixed crash when original layer has no `add_x/y` property
- `AddTilelayer()` can now add Tiled-default properties to layer
- Generated layers also have "Parallax Factor" and "Horizontal/Vertical Offset" set after processing, i.e. in-editor view synced with in-game view

--------------------------------------------------

The tool is still quite rough, and obviously I haven't tidied the code since the only primary user is me, but it's fully functional now, at least on the test level attached on Trello. I'm uploading in case you want to try for yourself.

To recap, it takes a specified layer as input, then spits out a new layer with all the relevant properties assigned, e.g. "Parallax Factor" for in-editor view & `scroll_x` for in-game view.